+++ lib/core/services/injection_container.dart
@@ -1,18 +1,15 @@
 import 'package:get_it/get_it.dart';
 
 import '../../feature/browse/data/datasources/browse_remote_datasource.dart';
-import '../../feature/browse/data/datasources/browse_remote_datasource_impl.dart';
-import '../../feature/browse/data/repositories/browse_repository_impl.dart';
+import '../../feature/browse/data/repositories_impl/browse_repository_impl.dart';
 import '../../feature/browse/domain/repositories/browse_repository.dart';
 import '../../feature/browse/domain/usecases/get_genres_usecase.dart';
 import '../../feature/browse/domain/usecases/get_movie_discover_usecase.dart';
 import '../../feature/browse/presentation/business_logic/bloc/browse_bloc.dart';
 import '../../feature/home/data/datasources/home_remote_datasource.dart';
-import '../../feature/home/data/datasources/home_remote_datasource_impl.dart';
 import '../../feature/home/data/datasources/movie_details_remote_datasource.dart';
-import '../../feature/home/data/datasources/movie_details_remote_datasource_impl.dart';
-import '../../feature/home/data/repositories/home_repository_impl.dart';
-import '../../feature/home/data/repositories/movie_details_repository_impl.dart';
+import '../../feature/home/data/repositories_impl/home_repository_impl.dart';
+import '../../feature/home/data/repositories_impl/movie_details_repository_impl.dart';
 import '../../feature/home/domain/repositories/home_repository.dart';
 import '../../feature/home/domain/repositories/movie_details_repository.dart';
 import '../../feature/home/domain/usecases/get_details_usecase.dart';
@@ -24,8 +21,7 @@ import '../../feature/home/domain/usecases/get_upcoming_usecase.dart';
 import '../../feature/home/presentation/business_logic/bloc/home_screen_bloc.dart';
 import '../../feature/home/presentation/business_logic/bloc/movie_details_bloc.dart';
 import '../../feature/bookmarks/data/datasources/bookmark_remote_datasource.dart';
-import '../../feature/bookmarks/data/datasources/bookmark_remote_datasource_impl.dart';
-import '../../feature/bookmarks/data/repositories/bookmark_repository_impl.dart';
+import '../../feature/bookmarks/data/repositories_impl/bookmark_repository_impl.dart';
 import '../../feature/bookmarks/domain/repositories/bookmark_repository.dart';
 import '../../feature/bookmarks/domain/usecases/get_favorites_usecase.dart';
 import '../../feature/bookmarks/domain/usecases/add_favorite_usecase.dart';
@@ -33,13 +29,12 @@ import '../../feature/bookmarks/domain/usecases/delete_favorite_usecase.dart';
 import '../../feature/bookmarks/presentation/business_logic/bloc/bookmark_bloc.dart';
 import '../../feature/home/presentation/business_logic/cubit/bottom_nav_cubit.dart';
 import '../../feature/search/data/datasources/search_remote_datasource.dart';
-import '../../feature/search/data/datasources/search_remote_datasource_impl.dart';
-import '../../feature/search/data/repositories/search_repository_impl.dart';
+import '../../feature/search/data/repositories_impl/search_repository_impl.dart';
 import '../../feature/search/domain/repositories/search_repository.dart';
 import '../../feature/search/domain/usecases/search_movies_usecase.dart';
 import '../../feature/search/presentation/business_logic/bloc/search_bloc.dart';
 import '../network/remote/api_manager.dart';
-import '../network/dio_helper.dart';
+import '../network/remote/dio_helper.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import '../network/check_connectivity/cubit/connectivity_cubit.dart';
